Post rebase fixes
