+++ src/TestConditions.cpp
@@ -1,6 +1,5 @@
-#include "TestConditions.hpp"
-
-#define GAMBA_DEBUG true
+#include "headers/TestConditions.hpp"
+#include <cmath>
 
 namespace pde {
 
@@ -9,46 +8,6 @@ namespace pde {
 
 double TestNeumannConditions::parameter;
 
-namespace functions {
-
-// TODO: add explanation of functions
-
-/**
-* This is the function that we used until now to
-* pull the cube on x and y
-*/
-constexpr auto cube_pull = [](const Point<dim> &p) {
-    constexpr double small_tol = 1e-13;
-    const double tau = TestNeumannConditions::parameter;
-
-    if (std::abs(p[0]) < small_tol && std::abs(p[1]) > small_tol)
-	    return Point<dim>(-tau, 0, 0);
-    else if (std::abs(p[0]) > (1 - small_tol) && std::abs(p[1]) > small_tol)
-	    return Point<dim>(tau, 0, 0);
-    else if (std::abs(p[1]) < small_tol && std::abs(p[0]) > small_tol)
-	    return Point<dim>(0, -tau, 0);
-    else if (std::abs(p[1]) > (1 - small_tol) && std::abs(p[0]) > small_tol)
-	    return Point<dim>(0, tau, 0);
-    else
-	    return Point<dim>(0, 0, 0);
-};
-
-// FIXME: This function doesn't work
-// the simulation doesn't converge
-/**
-* This function should bend the rod as a horseshoe
-* it applies a force along z based on the x coordinate 
-* and proportional to tau
-*/
-constexpr auto rod_bend = [](const Point<dim> &p) {
-    const double tau = TestNeumannConditions::parameter;
-    // TODO: check that this function is in fact correct 
-    // and it's not the cause of the divergence
-    return Point<dim>(0, 0, tau * p[0]);
-};
-
-}
-
 // TODO: Implement this function
 
 const std::function<Point<dim> (const Point<dim> &)> 
@@ -56,14 +15,18 @@ const std::function<Point<dim> (const Point<dim> &)>
 {
     // Here I simply started defining different functions
     // The more models and boundary conditions we apply the better it is.
+    if(choice.starts_with("rod")) {
+    	return functions::rod_bend;
+    } 
+
+    if (choice.starts_with("bowl")) {
+	return functions::bowl_pull_out;
+    }
     if(choice.starts_with("cube")) {
 	return functions::cube_pull;
-    } else {
-	#if GAMBA_DEBUG
-	std::cout << "Calling rod bend" << std::endl;
-	#endif
-	return functions::rod_bend;
     }
+
+    throw std::runtime_error("Unknown Neumann Condition: " + choice);
 }
