+++ src/headers/Guccione.hpp
@@ -1,3 +1,6 @@
+#ifndef GUCCIONE_HPP
+#define GUCCIONE_HPP
+
 #include "MechanicalDisplacement.hpp"
 
 namespace pde {
@@ -19,14 +22,20 @@ public:
         const unsigned int &r_,
 	const std::map<types::boundary_id, const Function<dim> *> boundary_functions_,
         const std::function<Point<dim>(const Point<dim> &)> &neum_funcs_,
+        const std::unordered_set<int>& neumann_ids_,
 	const ForcingTermType forcing_term_,
+        const std::string& output_filename_,
+	const ConditionalOStream pcout_,
+	const unsigned int mpi_rank_,
 	const double param_c_,
         const std::array<double, 9> param_b_,
-        const std::function<std::array<Point<dim>, dim>(const Point<dim> &)> &aniso_fun_,
-        const std::unordered_set<int>& neumann_ids_,
-        const std::string& output_filename_
+        const std::function<std::array<Point<dim>, dim>(const Point<dim> &)> &aniso_fun_
 	):
-        MechanicalDisplacement(std::move(mesh_generator_), r_, boundary_functions_, neum_funcs_, forcing_term_, output_filename_),
+        MechanicalDisplacement(
+	    std::move(mesh_generator_), r_, 
+	    boundary_functions_, neum_funcs_, neumann_ids_,
+	    forcing_term_, output_filename_, pcout_, mpi_rank_
+	),
         param_c(param_c_),
         param_b(param_b_),
         aniso_fun(aniso_fun_)
@@ -38,3 +47,5 @@ public:
 };
 
 }
+
+#endif
