+++ src/main.cpp
@@ -1,207 +1,42 @@
-#include "NeoHooke.hpp"
-#include "Guccione.hpp"
-#include "TestConditions.hpp"
+#include "headers/NeoHooke.hpp"
+#include "headers/Guccione.hpp"
+#include "headers/TestConditions.hpp"
+#include "headers/Work.hpp"
 
+#include <filesystem>
 
 const pde::ForcingTermType select_forcing_term(std::string boolean_value) {
 	using namespace pde::TestForcingFunctions;
 	return std::stoi(boolean_value) ? bend_rod : null_forcing_term;
 }
 
-#include <iostream>
-#include <fstream>
-#include <optional>
-#include <vector>
-#include <string>
-
-template <int dim>
-class PullAcornOutwards
-{
-public:
-    PullAcornOutwards(double tau_)
-        : tau(tau_) {}
-
-    Point<dim> operator()(const Point<dim> &p) const {
-;;;;;;;;static const auto aa = 1 * 1, bb = 1 * 1, cc = 3 * 3;
-;;;;;;;;auto xx = p[0] * p[0], yy = p[1] * p[1], zz = p[2] * p[2];
-;;;;;;;;if (xx / aa + yy / bb + zz / cc > 0.9)
-;;;;;;;;;;;;return tau / std::sqrt(64 * xx + 64 * yy + 64 / 81 * zz) * Point<dim>(8 * p[0], 8 * p[1], 8 / 9 * p[2]); // vector normal to the outer surface of the acorn;
-;;;;;;;;return Point<dim>(0, 0, 0);
-    }
-
-private:
-    double tau = 0.5;
-};
-
-
-std::string trim(const std::string& s) {
-    auto start = s.find_first_not_of(" \t\r\n");
-    auto end   = s.find_last_not_of(" \t\r\n");
-    if (start == std::string::npos) return "";
-    return s.substr(start, end - start + 1);
-}
-
-std::unordered_set<int> parse_csv_ints(const std::string& s) {
-    std::unordered_set<int> out;
-    std::stringstream ss(s);
-    std::string token;
-    while (std::getline(ss, token, ',')) {
-        out.insert(std::stoi(trim(token)));
-    }
-    return out;
-}
-
-
-struct Work {
-    enum MaterialType { NeoHooke, Guccione };
-    MaterialType material;
-    enum GeometryType { File, Cube, Rod };
-    GeometryType geometry;
-    std::optional<std::string> filename; // only for file
-    std::string output_filename;
-
-    std::unordered_set<int> N_values;
-    std::string N_label;
-    std::string N_data;
-
-    struct DEntry {
-        int value;
-        // expand with different function on D boundary
-    };
-    std::vector<DEntry> D_entries;
-};
-
-std::vector<Work> parse_file(const std::string& path) {
-    std::ifstream in(path);
-    if (!in)
-        throw std::runtime_error("Cannot open file: " + path);
-
-    std::vector<Work> sections;
-    std::string line;
-
-    auto next_line = [&]() -> std::string {
-        if (!std::getline(in, line))
-            throw std::runtime_error("Unexpected end of file");
-        return trim(line);
-    };
-
-    while (std::getline(in, line)) {
-        line = trim(line);
-        if (line.empty())
-            continue;
-
-        if (line != "-----")
-            throw std::runtime_error("Expected section separator '-----'");
-
-        Work sec;
-
-        /* Material */
-        line = next_line();
-        if (line == "NeoHooke") sec.material = Work::MaterialType::NeoHooke;
-        else if (line == "Guccione") sec.material = Work::MaterialType::Guccione;
-        else throw std::runtime_error("Unknown material: " + line);
-
-        /* Geometry */
-        line = next_line();
-        {
-            std::stringstream ss(line);
-            std::string type;
-            ss >> type;
-
-            if (type == "file") {
-                sec.geometry = Work::GeometryType::File;
-                ss >> sec.filename.emplace();
-                if (!sec.filename || sec.filename->empty())
-                    throw std::runtime_error("Missing filename after 'file'");
-            }
-            else if (type == "cube") sec.geometry = Work::GeometryType::Cube;
-            else if (type == "rod")  sec.geometry = Work::GeometryType::Rod;
-            else throw std::runtime_error("Unknown geometry: " + type);
-        }
-
-        /* Output file */
-        line = next_line();
-        {
-            std::stringstream ss(line);
-            std::string word;
-            ss >> word;
-
-            if (word == "out") {
-                ss >> sec.output_filename;
-                if (sec.output_filename.empty())
-                    throw std::runtime_error("Missing filename after 'out'");
-            }
-            else throw std::runtime_error("Provide output filename as third line");
-        }
-
-        /* N line */
-        line = next_line();
-        if (line.empty() || line[0] != 'N')
-            throw std::runtime_error("Expected N line");
-
-        sec.N_values = parse_csv_ints(trim(line.substr(1)));
-
-        /* N data line */
-        line = next_line();
-        {
-            std::stringstream ss(line);
-            ss >> sec.N_label;
-            std::getline(ss, sec.N_data);
-            sec.N_data = trim(sec.N_data);
-        }
-
-        /* D lines */
-        while (true) {
-            std::streampos pos = in.tellg();
-            if (!std::getline(in, line))
-                break;
-
-            line = trim(line);
-            if (line.empty())
-                continue;
-
-            if (line == "-----") {
-                in.seekg(pos);
-                break;
-            }
-
-            if (line[0] != 'D')
-                throw std::runtime_error("Expected D line");
-
-            std::stringstream ss(line.substr(1));
-            int value;
-            std::string zero;
-
-            ss >> value >> zero;
-            if (zero != "zero")
-                throw std::runtime_error("Expected 'zero' after D value");
-
-            sec.D_entries.push_back({value});
-        }
-
-        sections.push_back(std::move(sec));
-    }
-
-    return sections;
-}
-
-
 int main(int argc, char *argv[])
 {
     using namespace pde;
-    const unsigned int dim = MechanicalDisplacement::dim;
+    constexpr unsigned int r = 1;
+    constexpr unsigned int C = 1; //Pa
+    constexpr unsigned int lambda = 2; //Pa
+
     Utilities::MPI::MPI_InitFinalize mpi_init(argc, argv);
 
-    if (argc < 2)
+    // This MPI process.
+    const unsigned int mpi_rank(Utilities::MPI::this_mpi_process(MPI_COMM_WORLD));
+    // Parallel output stream.
+    const ConditionalOStream pcout(std::cout, mpi_rank == 0);
+
+    if (argc < 3)
     {
-        std::cout << "Provide file with work\n";
+        pcout << "Provide file with work\n";
+	pcout << "Usage: ./PDE-06 <path/to/config_file.txt> [FORCING_TERM <0|1>]"
+			<< std::endl;
         return 1;
     }
-    
+
     std::string filename(argv[1]);
+
     if (!std::filesystem::exists(filename))
     {
-        std::cout << "File provided doesn't exist\n";
+        pcout << "File provided doesn't exist\n";
         return 1;
     }
 
@@ -209,79 +44,73 @@ int main(int argc, char *argv[])
     
     for (auto& w : work)
     {
-        std::cout << "Starting work:\n";
+        pcout << "Starting work:\n";
         std::unique_ptr<MeshGenerator<dim>> mesh_src;
         if (w.geometry == Work::GeometryType::File)
         {
-            std::cout << "Using mesh from file: " << w.filename.value() << '\n';
+            pcout << "Using mesh from file: " << w.filename.value() << '\n';
             mesh_src = std::make_unique<MeshLoader<dim>>(w.filename.value());
         }
         if (w.geometry == Work::GeometryType::Cube)
         {
-            std::cout << "Using cube mesh\n";
+            pcout << "Using cube mesh\n";
             mesh_src = std::make_unique<CubeGenerator<dim>>();
         }
         if (w.geometry == Work::GeometryType::Rod)
         {
-            std::cout << "Using rod mesh\n";
+            pcout << "Using rod mesh\n";
             mesh_src = std::make_unique<RodGenerator<dim>>();
         }
      
         std::map<types::boundary_id, const Function<dim>*> boundary_functions;
-        std::cout << "Diritto boundary on ids:";
+	Functions::ZeroFunction<dim> zero_function(dim);
+
+        pcout << "Diritto boundary on ids:";
         for (auto d : w.D_entries)
         {
-            std::cout << ' ' << d.value;
+            pcout << ' ' << d.value;
             boundary_functions[d.value] = &zero_function;
         }
-        std::cout << "\n";
-        std::cout << "UomoNuovo boundary on ids:";
+pcout << "\n";
+        pcout << "UomoNuovo boundary on ids:";
         for (auto d : w.N_values)
-            std::cout << ' ' << d;
-        std::cout << "\n";
+            pcout << ' ' << d;
+        pcout << "\n";
         
         std::function<Point<dim>(const Point<dim> &)> h;
-        if (w.N_label == "pull_out")
-        {
-            std::cout << "Neumann function PullAcornOutwards\n";
-            PullAcornOutwards<dim> fn(std::stod(w.N_data));
-            h = [fn](const Point<dim> &p) { return fn(p); };
-        }
-        else if (w.N_label == "pull_cube")
-        {
-            //Pulling the cube appart <-|  |-> in x,y directions
-            //The dirichlet conditions are set to 0 on the z boundaries
-            //The object can change volume, since lambda = 0 in our model
-            h = [](const Point<dim> &p) {
-                double tau_0 = 0.5;
-                double small_tol = 1e-13;
-                if (std::abs(p[0]) < small_tol && std::abs(p[1]) > small_tol)
-                    return Point<dim>(tau_0, 0, 0);
-                else if (std::abs(p[0]) > (1 - small_tol) && std::abs(p[1]) > small_tol)
-                    return Point<dim>(-tau_0, 0, 0);
-                else if (std::abs(p[1]) < small_tol && std::abs(p[0]) > small_tol)
-                    return Point<dim>(0, 0, 0);
-                else if (std::abs(p[1]) > (1 - small_tol) && std::abs(p[0]) > small_tol)
-                    return Point<dim>(0, 0, 0);
-                else
-                    return Point<dim>(0, 0, 0);
-            };
-        }
-        else
-        {
-            std::cout << "Unknown Neumann boundary function, skipping work\n";
+	try {
+	    if(w.N_data == "") {
+		pcout << "No parameter for Neumann Condition found.\n" <<
+			    "Falling back to standard: tau = 0.5" << std::endl;
+		TestNeumannConditions::initialize();
+	    } else {
+		TestNeumannConditions::initialize(std::stod(w.N_data));
+	    }
+	    h = TestNeumannConditions::choose_neumann_function(w.N_label);
+	} catch(std::invalid_argument &ia) {
+		pcout << "Invalid Argument: " << ia.what() << std::endl;
+	} catch(std::runtime_error &e) { 
+            pcout << e.what() << " skipping work" << std::endl;
             continue;
-        }
+	} 
 
+	// TODO: fix/add forcing term to work
         if (w.material == Work::MaterialType::NeoHooke)
         {
-            std::cout << "NeoHooke Problem\n";
-            NeoHooke problem = NeoHooke(std::move(mesh_src), r, boundary_functions, h, C, lambda, w.N_values, w.output_filename);
+            pcout << "NeoHooke Problem\n";
+            NeoHooke problem = NeoHooke(
+				std::move(mesh_src), r, 
+				boundary_functions, h, 
+				w.N_values, select_forcing_term(argv[2]),
+				w.output_filename,
+				pcout, mpi_rank,
+				C, lambda
+			);
             problem.setup();
             problem.solve();
             problem.output();
 
-            std::cout << "\n\n\n\n";
+            pcout << "\n\n\n\n";
         }
         else
         {
